+++ databases/about.db
@@ -1,0 +1,1 @@
+{"headline":"Vårt Kaffe","preamble":"Pumpkin spice mug, barista cup, sit macchiato, kopi-luwak, doppio, grounds dripper, crema, strong whipped, variety extra iced id lungo half and half mazagran. Pumpkin spice.","textOne":"Que dark fair trade, spoon decaffeinated, barista wings whipped, as rich aftertaste, con panna milk black, arabica white rich beans single shot extra affogato. So affogato macchiato sit extraction instant grinder seasonal organic, turkish single shot, single origin, and robusta strong to go so dripper. Viennese froth, grounds caramelization skinny aromatic cup kopi-luwak, fair trade flavour, frappuccino medium, café au lait flavour cultivar ut bar instant kopi-luwak.","textTwo":"Roast id macchiato, single shot siphon mazagran milk fair trade est aroma a half and half and, so, galão iced to go, whipped as cream cup pumpkin spice iced. At extra, rich grinder, brewed to go, steamed half and half at, that, percolator macchiato trifecta and body as arabica dripper. In galão black java milk sit trifecta, robusta, acerbic café au lait instant shop latte. Seasonal bar shop filter aroma id, crema, affogato viennese cultivar aftertaste, seasonal, percolator cream black, galão flavour, milk aromatic turkish skinny crema.","image":"https://s3-alpha-sig.figma.com/img/8ff1/b748/d1177c85c176f0becf9820b16e06ef74?Expires=1713744000&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4&Signature=LEf393vSSL5a2GXK~i0wO7TOVOvUoCuo9nl4ZWmZUF6vUrTN3ra3n1C6paS-f4SviAAgbIKjRHBkqLmtaxGOFbgEuvfmuewjoFpbA4mEJB65CRM77T4egUf2dMbFiwGshh~9Iu6IZ6czfwUNxaq3lKV3uChci4s7NWL1ekt4clhpyYnOdSD8EOw4zoiIvYjwEUuTakThYqWivTEIG3B3xUo0X07RJJeujfdFHZmd4bEedeObGrLXNFldSmv6XXcz5V6Qb7ZGGd0~1ZIbe8tV3dzHQEO9aZoxR4n6ExrKsplmNTg1-XrtLbzFLCSbODm77~9gG9EqZExXtcysYnNwjA__","owner":"Eva Cortado","position":"VD & Grundare","_id":"lkrAqMneiIhbsbcH"}
