+++ databases/menu.db
@@ -1,0 +1,6 @@
+{"title":"Cortado","desc":"Bryggd på månadens bönor.","price":39,"_id":"03YTBIVBS2bPgwZN"}
+{"title":"Kaffe Latte","desc":"Bryggd på månadens bönor.","price":54,"_id":"1R8zFT7MtuerQeXR"}
+{"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39,"_id":"9KDPg1C3kjdGdLy2"}
+{"title":"Caffè Doppio","desc":"Bryggd på månadens bönor.","price":49,"_id":"C3aJhbioyUDHylXZ"}
+{"title":"Cappuccino","desc":"Bryggd på månadens bönor.","price":49,"_id":"HxVmHFZrQ7fksh9v"}
+{"title":"Latte Macchiato","desc":"Bryggd på månadens bönor.","price":49,"_id":"Ii6Ic92URPql5CMN"}
