+++ databases/users.db
@@ -1,0 +1,2 @@
+{"username":"Egil","email":"egil@egilsson.se","password":"egil","userid":"476d0f6c-bf79-4a7d-bc43-5510b5160dc3","_id":"HDLRZM4KwWGJN7nc"}
+{"username":"Egilina","email":"egilina@egildottir.se","password":"egilina","userid":"bcbe96fe-56f4-490a-9b69-caead40c7502","_id":"SKyeIpH1ssPq3MWo"}
