--- meny.db
@@ -1,12 +1,0 @@
-{"title":"Napoleonbakelse","desc":"goa grejer","_id":"MMgWVaF35Tlz6ApT"}
-{"title":"Kaffe Latte","desc":"Bryggd på månadens bönor.","price":54,"_id":"MORJUtgUxD9EaqN7"}
-{"title":"Cookie","desc":"goa grejer","_id":"ONq5gnjb666z6SVh"}
-{"title":"Dadelboll","desc":"goa grejer","_id":"XKWJIW60HNAm8Xge"}
-{"title":"Caffè Doppio","desc":"Bryggd på månadens bönor.","price":49,"_id":"YlL8xiXmiSwj3UMS"}
-{"title":"Chokladboll","desc":"goa grejer","price":29,"_id":"aPoqEZL3o2j0FU7w"}
-{"title":"Cappuccino","desc":"Bryggd på månadens bönor.","price":49,"_id":"amFzVojTlweyPKc8"}
-{"title":"Cortado","desc":"Bryggd på månadens bönor.","price":39,"_id":"b84Lt54FjAmn3ZsI"}
-{"title":"kanelbulle","desc":"goa grejer","_id":"cJZ5fmZpAYBQJpQr"}
-{"title":"Latte Macchiato","desc":"Bryggd på månadens bönor.","price":49,"_id":"jvHNN0oeN7b0Yauo"}
-{"title":"Morotskaka","desc":"goa grejer","price":29,"_id":"oqXcpOGB43QjkuMP"}
-{"title":"Bryggkaffe","desc":"Bryggd på månadens bönor.","price":39,"_id":"zMowJhiTkwSDKNJU"}
