--- users.db
@@ -1,4 +1,0 @@
-{"username":"Gurra","email":"knugen@my.mail","password":"hundhäst","userid":"c7a61942-d75d-4153-901f-e26c135753ec","_id":"Emmf53r3jdJq7T2t"}
-{"username":"Egil","email":"egil@egilsson.nu","password":"egil","userid":"778eb471-2995-448f-93be-7ad725fc9adb","_id":"KJksSN2ebZVXvK5p"}
-{"username":"Bertil","email":"bertil@my.mail","password":"gethund","userid":"c9c7771b-1475-4864-8456-1935f3175ee0","_id":"VxthIZqmsbsBvs24"}
-{"username":"Klas","email":"Klättermus","password":"mjau","userid":"0e417a62-f3d3-42a8-ab50-ee455106e374","_id":"RfnZ8CuwGjyNfIAY"}
